requirements.txt
flask==3.0.3
ezdxf==1.3.4
supabase==2.10.0
httpx==0.27.2
python-dotenv==1.0.1
gunicorn==22.0.0
